Build this test with clang for ASAN_ALL.
Since some of the test's dependencies use clang, ASAN_ALL will cause
those to be built with ASAN instrumentation, meaning the final
executable needs to be linked with ASAN's runtime libraries.

The simple fix is to just move the test to clang to let it link with
ASAN when ASAN_ALL is provided.

Change-Id: I8d134667347ec45fe59d557e3c792520eaf62f1b